main: blink led wakeup pattern after initializing the USB Stack
Blinking the LEDs for each channel can add up to quite a significant
delay in the initialization, especially when initializing multiple CAN
channels, but even for single-channel devices sometimes kernel errors
are found on the host, since the host will try re-enumeration the usb
device in such case.

For devices that cannot request USB enumeration themselves due to
hardwired pullups on the USB D+ line this may lead to timeouts and errors
(Device not responding to setup address).

Blinking the LEDs after setting up the USB Peripheral and USB stack
fixes this issue.

Signed-off-by: Johannes Zink <opensource@johannes-zink.de>